+++ code/http4s/src/main/scala/com/avast/datadog4s/extension/http4s/impl/package.scala
@@ -1,0 +1,5 @@
+package com.avast.datadog4s.extension.http4s
+
+package object impl {
+  private[http4s] type ActiveConnections = Map[Option[String], Int]
+}